requirements.txt
# Let pip auto-resolve compatible versions
langchain
langgraph
langchain-openai
langchain-community
langchain-chroma

chromadb
PyPDF2
tiktoken
python-dotenv
pydantic
pyyaml
numpy
streamlit
langsmith
scikit-learn
pytesseract
pdf2image
Pillow
